+++ pcs/lib/corosync/config_validators.py
@@ -41,8 +41,10 @@ _QDEVICE_NET_REQUIRED_OPTIONS = (
 _QDEVICE_NET_OPTIONAL_OPTIONS = (
     "connect_timeout",
     "force_ip_version",
+    "keep_active_partition_tie_breaker",
     "port",
     "tie_breaker",
+    "tls",
 )
 
 
@@ -2110,6 +2112,12 @@ def _get_qdevice_model_net_options_validators(
         validate.ValueIn(
             "tie_breaker",
             ["lowest", "highest"] + sorted(node_ids),
+            severity=severity,
+        ),
+        validate.ValueIn("tls", ("on", "off", "required"), severity=severity),
+        validate.ValueIn(
+            "keep_active_partition_tie_breaker",
+            ("on", "off"),
             severity=severity,
         ),
     ]